@@ -27,19 +27,30 @@
   "prefer-stable": true,
   "minimum-stability": "dev",
   "require": {
-    "php": "^7.1",
+    "php": "^8.0|^8.1|^8.2|^8.3|^8.4",
     "ext-ctype": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "^7.5"
+    "dg/bypass-finals": "^1.9",
+    "phpunit/phpunit": "^9.0"
   },
   "autoload": {
     "psr-4": {
       "Jojo1981\\JsonAstBuilder\\": "src/"
     }
   },
+  "autoload-dev": {
+    "psr-4": {
+      "tests\\Jojo1981\\JsonAstBuilder\\": "tests/"
+    }
+  },
   "config": {
     "bin-dir": "bin",
     "sort-packages": true
+  },
+  "scripts": {
+    "test": [
+      "./bin/phpunit"
+    ]
   }
 }
