@@ -7,33 +7,52 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Ast;
 
 use Jojo1981\JsonAstBuilder\Visitor\VisitorInterface;
 
-class BooleanNode implements TypeNodeInterface
+/**
+ * @package Jojo1981\JsonAstBuilder\Ast
+ */
+final class BooleanNode implements TypeNodeInterface
 {
     use TokenAwareTrait;
 
     /** @var bool */
-    private $value;
+    private bool $value;
 
+    /**
+     * @param bool $value
+     */
     public function __construct(bool $value)
     {
         $this->value = $value;
     }
 
+    /**
+     * @return bool
+     */
     public function getValue(): bool
     {
         return $this->value;
     }
 
+    /**
+     * @param bool $value
+     * @return void
+     */
     public function setValue(bool $value): void
     {
         $this->value = $value;
     }
 
-    public function accept(VisitorInterface $visitor)
+    /**
+     * @param VisitorInterface $visitor
+     * @return mixed
+     */
+    public function accept(VisitorInterface $visitor): mixed
     {
         return $visitor->visitBooleanNode($this);
     }
