@@ -7,33 +7,49 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Ast;
 
 use Jojo1981\JsonAstBuilder\Visitor\VisitorInterface;
 
-class IntegerNode implements TypeNodeInterface
+/**
+ * @package Jojo1981\JsonAstBuilder\Ast
+ */
+final class IntegerNode implements TypeNodeInterface
 {
     use TokenAwareTrait;
 
     /** @var int */
-    private $value;
+    private int $value;
 
     public function __construct(int $value)
     {
         $this->value = $value;
     }
 
+    /**
+     * @return int
+     */
     public function getValue(): int
     {
         return $this->value;
     }
 
+    /**
+     * @param int $value
+     * @return void
+     */
     public function setValue(int $value): void
     {
         $this->value = $value;
     }
 
-    public function accept(VisitorInterface $visitor)
+    /**
+     * @param VisitorInterface $visitor
+     * @return mixed
+     */
+    public function accept(VisitorInterface $visitor): mixed
     {
         return $visitor->visitIntegerNode($this);
     }
