@@ -7,33 +7,52 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Ast;
 
 use Jojo1981\JsonAstBuilder\Visitor\VisitorInterface;
 
-class KeyNode implements NodeInterface
+/**
+ * @package Jojo1981\JsonAstBuilder\Ast
+ */
+final class KeyNode implements NodeInterface, NodeWithTokenInterface
 {
     use TokenAwareTrait;
 
     /** @var string */
-    private $value;
+    private string $value;
 
+    /**
+     * @param string $value
+     */
     public function __construct(string $value)
     {
         $this->value = $value;
     }
 
+    /**
+     * @return string
+     */
     public function getValue(): string
     {
         return $this->value;
     }
 
+    /**
+     * @param string $value
+     * @return void
+     */
     public function setValue(string $value): void
     {
         $this->value = $value;
     }
 
-    public function accept(VisitorInterface $visitor)
+    /**
+     * @param VisitorInterface $visitor
+     * @return mixed
+     */
+    public function accept(VisitorInterface $visitor): mixed
     {
         return $visitor->visitKeyNode($this);
     }
