@@ -7,33 +7,49 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Ast;
 
 use Jojo1981\JsonAstBuilder\Visitor\VisitorInterface;
 
-class NumberNode implements TypeNodeInterface
+/**
+ * @package Jojo1981\JsonAstBuilder\Ast
+ */
+final class NumberNode implements TypeNodeInterface
 {
     use TokenAwareTrait;
 
     /** @var float */
-    private $value;
+    private float $value;
 
     public function __construct(float $value)
     {
         $this->value = $value;
     }
 
+    /**
+     * @return float
+     */
     public function getValue(): float
     {
         return $this->value;
     }
 
+    /**
+     * @param float $value
+     * @return void
+     */
     public function setValue(float $value): void
     {
         $this->value = $value;
     }
 
-    public function accept(VisitorInterface $visitor)
+    /**
+     * @param VisitorInterface $visitor
+     * @return mixed
+     */
+    public function accept(VisitorInterface $visitor): mixed
     {
         return $visitor->visitNumberNode($this);
     }
