@@ -7,16 +7,22 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Ast;
 
 use Jojo1981\JsonAstBuilder\Visitor\VisitorInterface;
+use function array_walk;
 
-class ObjectNode implements TypeNodeInterface
+/**
+ * @package Jojo1981\JsonAstBuilder\Ast
+ */
+final class ObjectNode implements TypeNodeInterface
 {
     use TokenAwareTrait;
 
     /** @var MemberNode[] */
-    private $members = [];
+    private array $members = [];
 
     /**
      * @param MemberNode[] $members
@@ -41,7 +47,16 @@ class ObjectNode implements TypeNodeInterface
     public function setMembers(array $members): void
     {
         $this->members = [];
-        \array_walk($members, [$this, 'addMember']);
+        array_walk($members, [$this, 'addMember']);
+    }
+
+    /**
+     * @param VisitorInterface $visitor
+     * @return mixed
+     */
+    public function accept(VisitorInterface $visitor): mixed
+    {
+        return $visitor->visitObjectNode($this);
     }
 
     /**
@@ -51,10 +66,5 @@ class ObjectNode implements TypeNodeInterface
     private function addMember(MemberNode $member): void
     {
         $this->members[] = $member;
-    }
-
-    public function accept(VisitorInterface $visitor)
-    {
-        return $visitor->visitObjectNode($this);
     }
 }
