@@ -7,36 +7,51 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Lexer;
+
+use UnexpectedValueException;
+use function array_filter;
+use function array_map;
+use function array_merge;
+use function count;
+use function max;
+use function min;
+use function preg_match_all;
+use function sprintf;
+use function str_replace;
+use function strlen;
+use function substr;
 
 /**
  * @package Jojo1981\JsonAstBuilder\Lexer
  */
-class Scanner
+final class Scanner
 {
     /** @var string */
-    private $input;
+    private string $input;
 
     /** @var int */
-    private $length;
+    private int $length;
 
     /** @var bool */
-    private $empty;
+    private bool $empty;
 
     /** @var int */
-    private $lastIndex;
+    private int $lastIndex;
 
     /** @var int[] */
-    private $positionStops;
+    private array $positionStops;
 
     /** @var int */
-    private $position = 0;
+    private int $position = 0;
 
     /** @var int */
-    private $lineNumber = 0;
+    private int $lineNumber = 0;
 
     /** @var int */
-    private $linePosition = 0;
+    private int $linePosition = 0;
 
     /**
      * @param string $input
@@ -44,10 +59,56 @@ class Scanner
     public function __construct(string $input)
     {
         $this->input = $this->prepareInput($input);
-        $this->length = \strlen($this->input);
+        $this->length = strlen($this->input);
         $this->empty = 0 === $this->length;
         $this->lastIndex = $this->length - 1;
         $this->positionStops = $this->extractPositionStopsFromInput($this->input, $this->lastIndex);
+    }
+
+    /**
+     * Prepare input and convert all new line characters like carriage returns and line feed characters into the current
+     * system used end of line character.
+     *
+     * @param string $input
+     * @return string
+     */
+    private function prepareInput(string $input): string
+    {
+        $otherLineSeparators = array_filter(
+            ["\r\n", "\n\r", "\r", "\n"],
+            static function (string $lineSeparators): bool {
+                return PHP_EOL !== $lineSeparators;
+            }
+        );
+
+        return str_replace($otherLineSeparators, PHP_EOL, $input);
+    }
+
+    /**
+     * Extract all new line positions from the input. It will return an indexed array whose values are the found newline
+     * positions (zero based) in the input. The first position will be the start and the last position will be the end.
+     *
+     * @param string $input
+     * @param int $lastIndex
+     * @return int[]
+     */
+    private function extractPositionStopsFromInput(string $input, int $lastIndex): array
+    {
+        $result = [];
+        preg_match_all('/\n/', $input, $matches, PREG_OFFSET_CAPTURE);
+        if (count($matches[0])) {
+            $result = array_merge(
+                $result,
+                array_map(
+                    static function (array $match): int {
+                        return $match[1];
+                    },
+                    $matches[0]
+                )
+            );
+        }
+
+        return array_merge([0], $result, [$lastIndex]);
     }
 
     /**
@@ -64,21 +125,6 @@ class Scanner
     public function getPosition(): int
     {
         return $this->position;
-    }
-
-    /**
-     * @param int $length
-     * @throws \UnexpectedValueException
-     * @return null|string
-     */
-    public function look(int $length = 1): ?string
-    {
-        $this->assertPositiveInteger($length);
-        if ($this->hasEndReached() || $this->isEmpty()) {
-            return null;
-        }
-
-        return \substr($this->input, $this->position, $length);
     }
 
     /**
@@ -109,8 +155,8 @@ class Scanner
 
     /**
      * @param int $length
-     * @throws \UnexpectedValueException
-     * @return null|string
+     * @return string|null
+     * @throws UnexpectedValueException
      */
     public function read(int $length = 1): ?string
     {
@@ -124,19 +170,38 @@ class Scanner
     }
 
     /**
-     * @return bool
+     * @param int $integer
+     * @return void
+     * @throws UnexpectedValueException
      */
-    public function hasEndReached(): bool
+    private function assertPositiveInteger(int $integer): void
     {
-        return $this->position > $this->lastIndex;
+        if ($integer < 1) {
+            throw new UnexpectedValueException(sprintf('Expected a positive integer, but got: %d', $integer));
+        }
+    }
+
+    /**
+     * @param int $length
+     * @return string|null
+     * @throws UnexpectedValueException
+     */
+    public function look(int $length = 1): ?string
+    {
+        $this->assertPositiveInteger($length);
+        if ($this->hasEndReached() || $this->isEmpty()) {
+            return null;
+        }
+
+        return substr($this->input, $this->position, $length);
     }
 
     /**
      * @return bool
      */
-    public function isAtBegin(): bool
+    public function hasEndReached(): bool
     {
-        return 0 === $this->position;
+        return $this->position > $this->lastIndex;
     }
 
     /**
@@ -154,8 +219,8 @@ class Scanner
     private function movePosition(int $length): void
     {
         $newPosition = $this->position + $length;
-        $newPosition = $newPosition < 0 ? 0 : $newPosition;
-        $newPosition = ($newPosition > $this->length ? $this->length : $newPosition);
+        $newPosition = max($newPosition, 0);
+        $newPosition = (min($newPosition, $this->length));
 
         [$lineNumber, $linePosition] = $this->getLineNumberAndLinePositionForPosition($newPosition);
 
@@ -173,7 +238,7 @@ class Scanner
     private function getLineNumberAndLinePositionForPosition(int $position): array
     {
         $result = [0, 0];
-        $position = $position > $this->lastIndex ? $this->lastIndex : $position;
+        $position = min($position, $this->lastIndex);
 
         if ($position > 0) {
             foreach ($this->positionStops as $lineNumberZeroBased => $linePositionZeroBased) {
@@ -192,60 +257,10 @@ class Scanner
     }
 
     /**
-     * Prepare input and convert all new line characters like carriage returns and line feed characters into the current
-     * system used end of line character.
-     *
-     * @param string $input
-     * @return string
+     * @return bool
      */
-    private function prepareInput(string $input): string
+    public function isAtBegin(): bool
     {
-        $otherLineSeparators = \array_filter(
-            ["\r\n", "\n\r", "\r", "\n"],
-            static function (string $lineSeparators): bool {
-                return PHP_EOL !== $lineSeparators;
-            }
-        );
-
-        return \str_replace($otherLineSeparators, PHP_EOL, $input);
-    }
-
-    /**
-     * Extract all new line positions from the input. It will return an indexed array who's values are the found newline
-     * positions (zero based) in the input. The first position will be the start and the last position will be the end.
-     *
-     * @param string $input
-     * @param int $lastIndex
-     * @return int[]
-     */
-    private function extractPositionStopsFromInput(string $input, int $lastIndex): array
-    {
-        $result = [];
-        \preg_match_all('/\n/', $input, $matches, PREG_OFFSET_CAPTURE);
-        if (\count($matches[0])) {
-            $result = \array_merge(
-                $result,
-                \array_map(
-                    static function (array $match): int {
-                        return $match[1];
-                    },
-                    $matches[0]
-                )
-            );
-        }
-
-        return \array_merge([0], $result, [$lastIndex]);
-    }
-
-    /**
-     * @param int $integer
-     * @throws \UnexpectedValueException
-     * @return void
-     */
-    private function assertPositiveInteger(int $integer): void
-    {
-        if ($integer < 1) {
-            throw new \UnexpectedValueException(\sprintf('Expected a positive integer, but got: %d', $integer));
-        }
+        return 0 === $this->position;
     }
 }
