@@ -7,6 +7,8 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder;
 
 use Jojo1981\JsonAstBuilder\Ast\ArrayNode;
@@ -23,7 +25,15 @@ use Jojo1981\JsonAstBuilder\Ast\StringNode;
 use Jojo1981\JsonAstBuilder\Ast\ValueNode;
 use Jojo1981\JsonAstBuilder\Exception\ParseException;
 use Jojo1981\JsonAstBuilder\Lexer\LexerInterface;
+use Jojo1981\JsonAstBuilder\Lexer\Token;
 use Jojo1981\JsonAstBuilder\Lexer\TokenType;
+use UnexpectedValueException;
+use function chr;
+use function hexdec;
+use function in_array;
+use function preg_match;
+use function strlen;
+use function substr;
 
 /**
  * The parser is responsible for generating an AST from the tokens it will get from the lexer and performs the
@@ -31,7 +41,7 @@ use Jojo1981\JsonAstBuilder\Lexer\TokenType;
  *
  * @package Jojo1981\JsonAstBuilder
  */
-class Parser
+final class Parser
 {
     /** @var int[] */
     private const WHITE_SPACE_TOKENS = [
@@ -40,7 +50,7 @@ class Parser
     ];
 
     /** @var LexerInterface */
-    private $lexer;
+    private LexerInterface $lexer;
 
     /**
      * @param LexerInterface $lexer
@@ -52,8 +62,8 @@ class Parser
 
     /**
      * @param string $input
-     * @throws ParseException
      * @return void
+     * @throws ParseException
      */
     public function setInput(string $input): void
     {
@@ -61,9 +71,9 @@ class Parser
     }
 
     /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
      * @return JsonNode
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     public function parse(): JsonNode
     {
@@ -73,9 +83,9 @@ class Parser
     }
 
     /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
      * @return JsonNode
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     private function json(): JsonNode
     {
@@ -86,9 +96,9 @@ class Parser
     }
 
     /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
      * @return ElementNode
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     private function element(): ElementNode
     {
@@ -100,9 +110,21 @@ class Parser
     }
 
     /**
+     * @return void
+     * @throws UnexpectedValueException
      * @throws ParseException
-     * @throws \UnexpectedValueException
+     */
+    private function eatWhiteSpace(): void
+    {
+        while (in_array($this->lexer->getCurrent()->getType(), self::WHITE_SPACE_TOKENS, true)) {
+            $this->lexer->getNext();
+        }
+    }
+
+    /**
      * @return ValueNode
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     private function value(): ValueNode
     {
@@ -117,40 +139,48 @@ class Parser
 
         $currentToken = $this->lexer->getCurrent();
         switch ($currentToken->getType()) {
-            case (TokenType::TOKEN_EOF): {
+            case (TokenType::TOKEN_EOF):
+            {
                 throw ParseException::unexpectedEndOfFile($this->lexer->getCurrent(), $expectedTokens);
             }
-            case (TokenType::TOKEN_LEFT_CURLY_BRACKET): {
+            case (TokenType::TOKEN_LEFT_CURLY_BRACKET):
+            {
                 $typeNode = $this->object();
                 break;
             }
-            case (TokenType::TOKEN_LEFT_SQUARE_BRACKET): {
+            case (TokenType::TOKEN_LEFT_SQUARE_BRACKET):
+            {
                 $typeNode = $this->array();
                 break;
             }
-            case (TokenType::TOKEN_STRING): {
+            case (TokenType::TOKEN_STRING):
+            {
                 $typeNode = $this->string();
                 break;
             }
-            case (TokenType::TOKEN_NUMBER): {
+            case (TokenType::TOKEN_NUMBER):
+            {
                 $typeNode = new NumberNode((float) $currentToken->getLexeme());
                 $typeNode->setToken($currentToken);
                 $this->eatToken(TokenType::TOKEN_NUMBER);
                 break;
             }
-            case (TokenType::TOKEN_INT): {
+            case (TokenType::TOKEN_INT):
+            {
                 $typeNode = new IntegerNode((int) $currentToken->getLexeme());
                 $typeNode->setToken($currentToken);
                 $this->eatToken(TokenType::TOKEN_INT);
                 break;
             }
-            case (TokenType::TOKEN_KEYWORD && \in_array($currentToken->getLexeme(), ['true', 'false'], true)): {
+            case (TokenType::TOKEN_KEYWORD && in_array($currentToken->getLexeme(), ['true', 'false'], true)):
+            {
                 $typeNode = new BooleanNode('true' === $currentToken->getLexeme());
                 $typeNode->setToken($currentToken);
                 $this->eatToken(TokenType::TOKEN_KEYWORD);
                 break;
             }
-            case (TokenType::TOKEN_KEYWORD && 'null' === $currentToken->getLexeme()): {
+            case (TokenType::TOKEN_KEYWORD && 'null' === $currentToken->getLexeme()):
+            {
                 $typeNode = new NullNode();
                 $typeNode->setToken($currentToken);
                 $this->eatToken(TokenType::TOKEN_KEYWORD);
@@ -164,41 +194,9 @@ class Parser
     }
 
     /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
-     * @return StringNode
-     */
-    private function string(): StringNode
-    {
-        $currentToken = $this->lexer->getCurrent();
-        $this->eatToken(TokenType::TOKEN_STRING);
-
-        $stringNode = new StringNode(\substr($currentToken->getLexeme(), 1, -1));
-        $stringNode->setToken($currentToken);
-
-        return $stringNode;
-    }
-
-    /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
-     * @return KeyNode
-     */
-    private function key(): KeyNode
-    {
-        $currentToken = $this->lexer->getCurrent();
-        $this->eatToken(TokenType::TOKEN_STRING);
-
-        $keyNode = new KeyNode(\substr($currentToken->getLexeme(), 1, -1));
-        $keyNode->setToken($currentToken);
-
-        return $keyNode;
-    }
-
-    /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
      * @return ObjectNode
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     private function object(): ObjectNode
     {
@@ -216,9 +214,25 @@ class Parser
     }
 
     /**
+     * @param int $tokenType
+     * @return void
+     * @throws UnexpectedValueException
      * @throws ParseException
-     * @throws \UnexpectedValueException
+     */
+    private function eatToken(int $tokenType): void
+    {
+        if ($this->lexer->getCurrent()->getType() !== $tokenType) {
+            throw ParseException::unexpectedToken($this->lexer->getCurrent(), [TokenType::getLiteral($tokenType)]);
+        }
+        if (TokenType::TOKEN_EOF !== $tokenType) {
+            $this->lexer->getNext();
+        }
+    }
+
+    /**
      * @return MemberNode[]
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     private function members(): array
     {
@@ -239,9 +253,9 @@ class Parser
     }
 
     /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
      * @return MemberNode
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     private function member(): MemberNode
     {
@@ -255,9 +269,25 @@ class Parser
     }
 
     /**
+     * @return KeyNode
+     * @throws UnexpectedValueException
      * @throws ParseException
-     * @throws \UnexpectedValueException
+     */
+    private function key(): KeyNode
+    {
+        $currentToken = $this->lexer->getCurrent();
+        $this->eatToken(TokenType::TOKEN_STRING);
+
+        $keyNode = new KeyNode(self::parseStringValue($currentToken));
+        $keyNode->setToken($currentToken);
+
+        return $keyNode;
+    }
+
+    /**
      * @return ArrayNode
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     private function array(): ArrayNode
     {
@@ -274,9 +304,9 @@ class Parser
     }
 
     /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
      * @return ElementNode[]
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     private function elements(): array
     {
@@ -297,30 +327,104 @@ class Parser
     }
 
     /**
+     * @return StringNode
+     * @throws UnexpectedValueException
      * @throws ParseException
-     * @throws \UnexpectedValueException
-     * @return void
      */
-    private function eatWhiteSpace(): void
+    private function string(): StringNode
     {
-        while (\in_array($this->lexer->getCurrent()->getType(), self::WHITE_SPACE_TOKENS, true)) {
-            $this->lexer->getNext();
-        }
+        $currentToken = $this->lexer->getCurrent();
+        $this->eatToken(TokenType::TOKEN_STRING);
+
+        $stringNode = new StringNode(self::parseStringValue($currentToken));
+        $stringNode->setToken($currentToken);
+
+        return $stringNode;
     }
 
     /**
-     * @param int $tokenType
-     * @throws ParseException
-     * @throws \UnexpectedValueException
-     * @return void
+     * @param Token $token
+     * @return string
+     * @throws UnexpectedValueException
      */
-    private function eatToken(int $tokenType): void
+    private static function parseStringValue(Token $token): string
     {
-        if ($this->lexer->getCurrent()->getType() !== $tokenType) {
-            throw ParseException::unexpectedToken($this->lexer->getCurrent(), [TokenType::getLiteral($tokenType)]);
+        if ($token->getType() !== TokenType::TOKEN_STRING) {
+            throw new UnexpectedValueException('Token is not a string');
         }
-        if (TokenType::TOKEN_EOF !== $tokenType) {
-            $this->lexer->getNext();
+
+        $result = '';
+        $rawString = substr($token->getLexeme(), 1, -1);
+        $length = strlen($rawString);
+        for ($i = 0; $i < $length; $i++) {
+            $char = $rawString[$i];
+            if ('\\' === $char) {
+                $nextChar = $rawString[$i + 1] ?? '';
+                if (in_array($nextChar, ['"', '\\', '/', 'b', 'f', 'n', 'r', 't'], true)) {
+                    $result .= self::parseControlChar($char . $nextChar);
+                    $i++;
+                } elseif ('u' === $nextChar) {
+                    $unicodeSequence = substr($rawString, $i, 6);
+                    if (strlen($unicodeSequence) !== 6 || !preg_match('/\\\\u[0-9a-fA-F]{4}/', $unicodeSequence)) {
+                        throw new UnexpectedValueException('Invalid Unicode escape sequence: ' . $unicodeSequence);
+                    }
+                    $result .= self::parseUnicode($unicodeSequence);
+                    $i += 5;
+                } else {
+                    throw new UnexpectedValueException('Invalid escape sequence: \\' . $nextChar);
+                }
+            } else {
+                $result .= $char;
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param string $sequence
+     * @return string
+     * @throws UnexpectedValueException
+     */
+    private static function parseControlChar(string $sequence): string
+    {
+        return match ($sequence) {
+            '\\"' => '"',
+            '\\\\' => '\\',
+            '\\/' => '/',
+            '\\b' => "\b",
+            '\\f' => "\f",
+            '\\n' => "\n",
+            '\\r' => "\r",
+            '\\t' => "\t",
+            default => throw new UnexpectedValueException('Invalid escape sequence: ' . $sequence)
+        };
+    }
+
+    /**
+     * @param string $sequence
+     * @return string
+     * @throws UnexpectedValueException
+     */
+    private static function parseUnicode(string $sequence): string
+    {
+        $codePoint = hexdec(substr($sequence, 2));
+        if ($codePoint <= 0x7F) {
+            return chr($codePoint);
+        } elseif ($codePoint <= 0x7FF) {
+            return chr(0xC0 | ($codePoint >> 6)) .
+                   chr(0x80 | ($codePoint & 0x3F));
+        } elseif ($codePoint <= 0xFFFF) {
+            return chr(0xE0 | ($codePoint >> 12)) .
+                   chr(0x80 | (($codePoint >> 6) & 0x3F)) .
+                   chr(0x80 | ($codePoint & 0x3F));
+        } elseif ($codePoint <= 0x10FFFF) {
+            return chr(0xF0 | ($codePoint >> 18)) .
+                   chr(0x80 | (($codePoint >> 12) & 0x3F)) .
+                   chr(0x80 | (($codePoint >> 6) & 0x3F)) .
+                   chr(0x80 | ($codePoint & 0x3F));
+        } else {
+            throw new UnexpectedValueException('Invalid Unicode code point: ' . $sequence);
         }
     }
 }
