@@ -7,6 +7,8 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Visitor;
 
 use Jojo1981\JsonAstBuilder\Ast\ArrayNode;
@@ -21,23 +23,26 @@ use Jojo1981\JsonAstBuilder\Ast\NumberNode;
 use Jojo1981\JsonAstBuilder\Ast\ObjectNode;
 use Jojo1981\JsonAstBuilder\Ast\StringNode;
 use Jojo1981\JsonAstBuilder\Ast\ValueNode;
+use function array_merge;
+use function count;
+use function str_repeat;
 
 /**
  * @package Jojo1981\JsonAstBuilder\Visitor
  */
-class JsonStringGeneratorVisitor implements VisitorInterface
+final class JsonStringGeneratorVisitor implements VisitorInterface
 {
     /** @var array */
-    private $options;
+    private array $options;
 
     /** @var string */
-    private $result = '';
+    private string $result = '';
 
     /** @var int */
-    private $level = 0;
+    private int $level = 0;
 
     /** @var string */
-    private $indent;
+    private string $indent;
 
     public function __construct(array $options = [])
     {
@@ -49,35 +54,54 @@ class JsonStringGeneratorVisitor implements VisitorInterface
             'spacesAfterColon' => 1,
             'lineSeparator' => PHP_EOL
         ];
-        $this->options = \array_merge($defaults, $options);
-        $this->indent = \str_repeat($this->options['useTabs'] ? "\t" : ' ', $this->options['indentSize']);
+        $this->options = array_merge($defaults, $options);
+        $this->indent = str_repeat($this->options['useTabs'] ? "\t" : ' ', $this->options['indentSize']);
     }
 
+    /**
+     * @return string
+     */
     public function getResult(): string
     {
         return $this->result;
     }
 
-    public function visitJsonNode(JsonNode $jsonNode): void
+    /**
+     * @param JsonNode $jsonNode
+     * @return mixed
+     */
+    public function visitJsonNode(JsonNode $jsonNode): mixed
     {
-        $jsonNode->getElement()->accept($this);
+        return $jsonNode->getElement()->accept($this);
     }
 
-    public function visitElementNode(ElementNode $elementNode): void
+    /**
+     * @param ElementNode $elementNode
+     * @return mixed
+     */
+    public function visitElementNode(ElementNode $elementNode): mixed
     {
-        $elementNode->getValue()->accept($this);
+        return $elementNode->getValue()->accept($this);
     }
 
-    public function visitValueNode(ValueNode $valueNode): void
+    /**
+     * @param ValueNode $valueNode
+     * @return mixed
+     */
+    public function visitValueNode(ValueNode $valueNode): mixed
     {
-        $valueNode->getType()->accept($this);
+        return $valueNode->getType()->accept($this);
     }
 
-    public function visitObjectNode(ObjectNode $objectNode): void
+    /**
+     * @param ObjectNode $objectNode
+     * @return mixed
+     */
+    public function visitObjectNode(ObjectNode $objectNode): mixed
     {
         $this->addText('{');
 
-        $memberCount = \count($objectNode->getMembers());
+        $memberCount = count($objectNode->getMembers());
         if ($memberCount > 0) {
             $this->level++;
             $this->addNewline();
@@ -97,13 +121,48 @@ class JsonStringGeneratorVisitor implements VisitorInterface
         }
 
         $this->addText('}');
+
+        return null;
     }
 
-    public function visitArrayNode(ArrayNode $arrayNode): void
+    /**
+     * @param string $text
+     * @return void
+     */
+    private function addText(string $text): void
+    {
+        $this->result .= $text;
+    }
+
+    /**
+     * @return void
+     */
+    private function addNewline(): void
+    {
+        if ($this->options['pretty']) {
+            $this->result .= $this->options['lineSeparator'];
+        }
+    }
+
+    /**
+     * @return void
+     */
+    private function addIndent(): void
+    {
+        if ($this->options['pretty']) {
+            $this->result .= str_repeat($this->indent, $this->level);
+        }
+    }
+
+    /**
+     * @param ArrayNode $arrayNode
+     * @return mixed
+     */
+    public function visitArrayNode(ArrayNode $arrayNode): mixed
     {
         $this->addText('[');
 
-        $elementCount = \count($arrayNode->getElements());
+        $elementCount = count($arrayNode->getElements());
         if ($elementCount > 0) {
             $this->level++;
             $this->addNewline();
@@ -123,89 +182,88 @@ class JsonStringGeneratorVisitor implements VisitorInterface
             $this->addIndent();
         }
         $this->addText(']');
+
+        return null;
     }
 
-    public function visitStringNode(StringNode $stringNode): void
+    /**
+     * @param StringNode $stringNode
+     * @return mixed
+     */
+    public function visitStringNode(StringNode $stringNode): mixed
     {
         $this->addText('"' . $stringNode->getValue() . '"');
+
+        return null;
     }
 
-    public function visitKeyNode(KeyNode $keyNode): void
+    /**
+     * @param KeyNode $keyNode
+     * @return mixed
+     */
+    public function visitKeyNode(KeyNode $keyNode): mixed
     {
         $this->addText('"' . $keyNode->getValue() . '"');
+
+        return null;
     }
 
-    public function visitNumberNode(NumberNode $numberNode): void
+    /**
+     * @param NumberNode $numberNode
+     * @return mixed
+     */
+    public function visitNumberNode(NumberNode $numberNode): mixed
     {
         $this->addText((string) $numberNode->getValue());
+
+        return null;
     }
 
     /**
      * @param IntegerNode $integerNode
-     * @return void
+     * @return mixed
      */
-    public function visitIntegerNode(IntegerNode $integerNode): void
+    public function visitIntegerNode(IntegerNode $integerNode): mixed
     {
         $this->addText((string) $integerNode->getValue());
+
+        return null;
     }
 
     /**
      * @param BooleanNode $booleanNode
-     * @return void
+     * @return mixed
      */
-    public function visitBooleanNode(BooleanNode $booleanNode): void
+    public function visitBooleanNode(BooleanNode $booleanNode): mixed
     {
         $this->addText($booleanNode->getValue() ? 'true' : 'false');
+
+        return null;
     }
 
     /**
      * @param NullNode $nullNode
-     * @return void
+     * @return mixed
      */
-    public function visitNullNode(NullNode $nullNode): void
+    public function visitNullNode(NullNode $nullNode): mixed
     {
         $this->addText('null');
+
+        return null;
     }
 
     /**
      * @param MemberNode $memberNode
-     * @return void
+     * @return mixed
      */
-    public function visitMemberNode(MemberNode $memberNode): void
+    public function visitMemberNode(MemberNode $memberNode): mixed
     {
         $memberNode->getKey()->accept($this);
-        $spaceBefore = \str_repeat(' ', $this->options['spacesBeforeColon']);
-        $spaceAfter = \str_repeat(' ', $this->options['spacesAfterColon']);
+        $spaceBefore = str_repeat(' ', $this->options['spacesBeforeColon']);
+        $spaceAfter = str_repeat(' ', $this->options['spacesAfterColon']);
         $this->addText($spaceBefore . ':' . $spaceAfter);
         $memberNode->getValue()->accept($this);
-    }
 
-    /**
-     * @return void
-     */
-    private function addNewline(): void
-    {
-        if ($this->options['pretty']) {
-            $this->result .= $this->options['lineSeparator'];
-        }
-    }
-
-    /**
-     * @return void
-     */
-    private function addIndent(): void
-    {
-        if ($this->options['pretty']) {
-            $this->result .= \str_repeat($this->indent, $this->level);
-        }
-    }
-
-    /**
-     * @param string $text
-     * @return void
-     */
-    private function addText(string $text): void
-    {
-        $this->result .= $text;
+        return null;
     }
 }
