@@ -7,6 +7,8 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Visitor;
 
 use Jojo1981\JsonAstBuilder\Ast\ArrayNode;
@@ -21,160 +23,52 @@ use Jojo1981\JsonAstBuilder\Ast\NumberNode;
 use Jojo1981\JsonAstBuilder\Ast\ObjectNode;
 use Jojo1981\JsonAstBuilder\Ast\StringNode;
 use Jojo1981\JsonAstBuilder\Ast\ValueNode;
+use Jojo1981\JsonAstBuilder\Helper\PlantUmlHelper;
+use function array_key_exists;
+use function array_pop;
+use function end;
+use function explode;
+use function get_class;
 
 /**
  * @package Jojo1981\JsonAstBuilder\Visitor
  */
-class PlantUmlAstNodesGeneratorVisitor implements VisitorInterface
+final class PlantUmlAstNodesGeneratorVisitor implements VisitorInterface
 {
     /** @var string[] */
-    private $objects = [];
-
-    /** @var array  */
-    private $links = [];
-
-    /** @var int */
-    private $level = 0;
-
-    /** @var int[] */
-    private $nodeCounts = [];
+    private array $objects = [];
 
     /** @var string[] */
-    private $parents = [];
+    private array $links = [];
 
     /** @var int */
-    private $maxLevel = 20;
+    private int $level = 0;
+
+    /** @var int[] */
+    private array $nodeCounts = [];
+
+    /** @var string[] */
+    private array $parents = [];
+
+    /** @var int */
+    private int $maxLevel = 20;
 
     /**
      * @return string
      */
     public function getResult(): string
     {
-        $lines = [];
-        $lines[] = '@startuml';
-        $lines[] = '';
-        $lines[] = 'hide empty members';
-        $lines[] = '';
-        $lines[] = 'title';
-        $lines[] = 'AST objects diagram';
-        $lines[] = '';
-        $lines[] = 'end title';
-        $lines[] = '';
-        $lines[] = \implode(PHP_EOL, $this->objects);
-        $lines[] = '';
-        $lines[] = \implode(PHP_EOL, $this->links);
-        $lines[] = '@enduml';
-
-        return \implode(PHP_EOL, $lines);
+        return PlantUmlHelper::generateDocument($this->objects, $this->links);
     }
 
-    public function visitJsonNode(JsonNode $jsonNode): void
+    public function visitJsonNode(JsonNode $jsonNode): mixed
     {
         $this->addObjectForNode('JsonNode', '  element: ElementNode');
         $this->pushParent('JsonNode');
         $jsonNode->getElement()->accept($this);
         $this->popParent();
-    }
 
-    public function visitElementNode(ElementNode $elementNode): void
-    {
-        $this->addObjectForNode('ElementNode', '  value: ValueNode');
-
-        $this->pushParent('ElementNode');
-        $elementNode->getValue()->accept($this);
-        $this->popParent();
-
-    }
-
-    public function visitValueNode(ValueNode $valueNode): void
-    {
-        $parts = \explode('\\', \get_class($valueNode->getType()));
-
-        $this->addObjectForNode('ValueNode', '  type: ' . \end($parts));
-        $this->pushParent('ValueNode');
-        $valueNode->getType()->accept($this);
-        $this->popParent();
-
-    }
-
-    public function visitObjectNode(ObjectNode $objectNode): void
-    {
-        $this->addObjectForNode('ObjectNode', '  members: MemberNode[]');
-        $this->pushParent('ObjectNode');
-        foreach ($objectNode->getMembers() as $memberNode) {
-            $memberNode->accept($this);
-        }
-
-        $this->popParent();
-    }
-
-    /**
-     * @param MemberNode $memberNode
-     * @return void
-     */
-    public function visitMemberNode(MemberNode $memberNode): void
-    {
-        $this->addObjectForNode('MemberNode', '  key: KeyNode' . PHP_EOL . '  value: ElementNode');
-        $this->pushParent('MemberNode');
-        $memberNode->getKey()->accept($this);
-        $memberNode->getValue()->accept($this);
-        $this->popParent();
-    }
-
-    /**
-     * @param KeyNode $keyNode
-     * @return void
-     */
-    public function visitKeyNode(KeyNode $keyNode): void
-    {
-        $this->addObjectForNode('KeyNode', '  value: ' . $keyNode->getValue());
-    }
-
-    public function visitArrayNode(ArrayNode $arrayNode): void
-    {
-        $this->addObjectForNode('ArrayNode', '  elements: ElementNode[]');
-        $this->pushParent('ArrayNode');
-        foreach ($arrayNode->getElements() as $elementNode) {
-            $elementNode->accept($this);
-        }
-        $this->popParent();
-    }
-
-    public function visitStringNode(StringNode $stringNode): void
-    {
-        $this->addObjectForNode('StringNode', '  value: ' . $stringNode->getValue());
-    }
-
-    public function visitNumberNode(NumberNode $numberNode): void
-    {
-        $this->addObjectForNode('NumberNode', '  value: ' . $numberNode->getValue());
-    }
-
-    /**
-     * @param IntegerNode $integerNode
-     * @return void
-     */
-    public function visitIntegerNode(IntegerNode $integerNode): void
-    {
-        $this->addObjectForNode('IntegerNode', '  value: ' . $integerNode->getValue());
-    }
-
-    /**
-     * @param BooleanNode $booleanNode
-     * @return void
-     */
-    public function visitBooleanNode(BooleanNode $booleanNode): void
-    {
-        $this->addObjectForNode('BooleanNode', '  value: ' . ($booleanNode->getValue() ? 'true' : 'false'));
-    }
-
-    /**
-     * @param NullNode $nullNode
-     * @return void
-     */
-    public function visitNullNode(NullNode $nullNode): void
-    {
-        $this->addObjectForNode('NullNode');
+        return null;
     }
 
     /**
@@ -201,13 +95,44 @@ class PlantUmlAstNodesGeneratorVisitor implements VisitorInterface
     }
 
     /**
+     * @param string $nodeName
+     * @return void
+     */
+    private function incrementNodeCount(string $nodeName): void
+    {
+        if (!array_key_exists($nodeName, $this->nodeCounts)) {
+            $this->nodeCounts[$nodeName] = 0;
+        }
+
+        $this->nodeCounts[$nodeName]++;
+    }
+
+    /**
+     * @param string $nodeName
+     * @return string
+     */
+    private function getInstanceName(string $nodeName): string
+    {
+        if (0 === $this->level) {
+            return $nodeName;
+        }
+
+        $name = $nodeName . $this->level;
+        if (array_key_exists($nodeName, $this->nodeCounts)) {
+            $name .= $this->nodeCounts[$nodeName];
+        }
+
+        return $name;
+    }
+
+    /**
      * @param string $name
      * @return void
      */
     private function linkWithParent(string $name): void
     {
         if (!empty($this->parents) && $this->level < $this->maxLevel) {
-            $this->links[] = \end($this->parents) . ' --* ' . $name;
+            $this->links[] = end($this->parents) . ' --* ' . $name;
         }
     }
 
@@ -227,37 +152,125 @@ class PlantUmlAstNodesGeneratorVisitor implements VisitorInterface
     private function popParent(): void
     {
         $this->level--;
-        \array_pop($this->parents);
+        array_pop($this->parents);
+    }
+
+    public function visitElementNode(ElementNode $elementNode): mixed
+    {
+        $this->addObjectForNode('ElementNode', '  value: ValueNode');
+        $this->pushParent('ElementNode');
+        $elementNode->getValue()->accept($this);
+        $this->popParent();
+
+        return null;
+
+    }
+
+    public function visitValueNode(ValueNode $valueNode): mixed
+    {
+        $parts = explode('\\', get_class($valueNode->getType()));
+        $this->addObjectForNode('ValueNode', '  type: ' . end($parts));
+        $this->pushParent('ValueNode');
+        $valueNode->getType()->accept($this);
+        $this->popParent();
+
+        return null;
+    }
+
+    public function visitObjectNode(ObjectNode $objectNode): mixed
+    {
+        $this->addObjectForNode('ObjectNode', '  members: MemberNode[]');
+        $this->pushParent('ObjectNode');
+        foreach ($objectNode->getMembers() as $memberNode) {
+            $memberNode->accept($this);
+        }
+        $this->popParent();
+
+        return null;
     }
 
     /**
-     * @param string $nodeName
-     * @return void
+     * @param MemberNode $memberNode
+     * @return mixed
      */
-    private function incrementNodeCount(string $nodeName): void
+    public function visitMemberNode(MemberNode $memberNode): mixed
     {
-        if (!\array_key_exists($nodeName, $this->nodeCounts)) {
-            $this->nodeCounts[$nodeName] = 0;
-        }
+        $this->addObjectForNode('MemberNode', '  key: KeyNode' . PHP_EOL . '  value: ElementNode');
+        $this->pushParent('MemberNode');
+        $memberNode->getKey()->accept($this);
+        $memberNode->getValue()->accept($this);
+        $this->popParent();
 
-        $this->nodeCounts[$nodeName]++;
+        return null;
     }
 
     /**
-     * @param string $nodeName
-     * @return string
+     * @param KeyNode $keyNode
+     * @return mixed
      */
-    private function getInstanceName(string $nodeName): string
+    public function visitKeyNode(KeyNode $keyNode): mixed
     {
-        if (0 === $this->level) {
-            return $nodeName;
-        }
+        $this->addObjectForNode('KeyNode', '  value: ' . $keyNode->getValue());
 
-        $name = $nodeName . $this->level;
-        if (\array_key_exists($nodeName, $this->nodeCounts)) {
-            $name .= $this->nodeCounts[$nodeName];
-        }
+        return null;
+    }
 
-        return $name;
+    public function visitArrayNode(ArrayNode $arrayNode): mixed
+    {
+        $this->addObjectForNode('ArrayNode', '  elements: ElementNode[]');
+        $this->pushParent('ArrayNode');
+        foreach ($arrayNode->getElements() as $elementNode) {
+            $elementNode->accept($this);
+        }
+        $this->popParent();
+
+        return null;
+    }
+
+    public function visitStringNode(StringNode $stringNode): mixed
+    {
+        $this->addObjectForNode('StringNode', '  value: ' . $stringNode->getValue());
+
+        return null;
+    }
+
+    public function visitNumberNode(NumberNode $numberNode): mixed
+    {
+        $this->addObjectForNode('NumberNode', '  value: ' . $numberNode->getValue());
+
+        return null;
+    }
+
+    /**
+     * @param IntegerNode $integerNode
+     * @return mixed
+     */
+    public function visitIntegerNode(IntegerNode $integerNode): mixed
+    {
+        $this->addObjectForNode('IntegerNode', '  value: ' . $integerNode->getValue());
+
+        return null;
+    }
+
+    /**
+     * @param BooleanNode $booleanNode
+     * @return mixed
+     */
+    public function visitBooleanNode(BooleanNode $booleanNode): mixed
+    {
+        $this->addObjectForNode('BooleanNode', '  value: ' . ($booleanNode->getValue() ? 'true' : 'false'));
+
+        return null;
+    }
+
+    /**
+     * @param NullNode $nullNode
+     * @return mixed
+     */
+    public function visitNullNode(NullNode $nullNode): mixed
+    {
+        $this->addObjectForNode('NullNode');
+
+        return null;
     }
 }
