@@ -2,7 +2,7 @@
 /*
  * This file is part of the jojo1981/json-ast-builder package
  *
- * Copyright (c) 2019 Joost Nijhuis <jnijhuis81@gmail.com>
+ * Copyright (c) 2025 Joost Nijhuis <jnijhuis81@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
@@ -16,25 +16,16 @@ use Jojo1981\JsonAstBuilder\Lexer\Token;
 /**
  * @package Jojo1981\JsonAstBuilder\Ast
  */
-trait TokenAwareTrait
+interface NodeWithTokenInterface
 {
-    /** @var Token|null */
-    private ?Token $token = null;
-
     /**
      * @return Token|null
      */
-    public function getToken(): ?Token
-    {
-        return $this->token;
-    }
+    public function getToken(): ?Token;
 
     /**
      * @param Token|null $token
      * @return void
      */
-    public function setToken(?Token $token): void
-    {
-        $this->token = $token;
-    }
+    public function setToken(?Token $token): void;
 }
