@@ -7,8 +7,15 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Exception;
 
-class JsonException extends \DomainException
+use DomainException;
+
+/**
+ * @package Jojo1981\JsonAstBuilder\Exception
+ */
+abstract class JsonException extends DomainException
 {
 }
