@@ -7,14 +7,19 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Exception;
 
 use Jojo1981\JsonAstBuilder\Lexer\Token;
+use function count;
+use function implode;
+use function sprintf;
 
 /**
  * @package Jojo1981\JsonAstBuilder\Exception
  */
-class ParseException extends JsonException
+final class ParseException extends JsonException
 {
     /**
      * @param string $char
@@ -28,9 +33,8 @@ class ParseException extends JsonException
         int $position,
         int $lineNumber,
         int $linePosition
-    ): ParseException
-    {
-        return new static(\sprintf(
+    ): ParseException {
+        return new self(sprintf(
             'Unknown character: `%s` found at position: %d [%d:%d].',
             $char,
             $position,
@@ -47,7 +51,7 @@ class ParseException extends JsonException
      */
     public static function unexpectedEndOfFileFound(int $position, int $lineNumber, int $linePosition): ParseException
     {
-        return new static(\sprintf(
+        return new self(sprintf(
             'Unexpected end of file reached at position: %d [%d:%d]',
             $position,
             $lineNumber,
@@ -60,7 +64,7 @@ class ParseException extends JsonException
      */
     public static function invalidInput(): ParseException
     {
-        return new static('Empty input given which is invalid json content');
+        return new self('Empty input given which is invalid json content');
     }
 
     /**
@@ -77,9 +81,8 @@ class ParseException extends JsonException
         int $lineNumber,
         int $linePosition,
         string $expectedKeyWord
-    ): ParseException
-    {
-        return new static(\sprintf(
+    ): ParseException {
+        return new self(sprintf(
             'Invalid characters: `%s` found at position: %d [%d:%d]. Did you mean: `%s`?',
             $chars,
             $position,
@@ -97,7 +100,7 @@ class ParseException extends JsonException
      */
     public static function unterminatedStringFound(int $position, int $lineNumber, int $linePosition): ParseException
     {
-        return new static(\sprintf(
+        return new self(sprintf(
             'Unterminated string found at position: %d [%d:%d]',
             $position,
             $lineNumber,
@@ -114,7 +117,7 @@ class ParseException extends JsonException
      */
     public static function illegalCharacterInStringFound(int $position, int $lineNumber, int $linePosition, string $char): ParseException
     {
-        return new static(\sprintf(
+        return new self(sprintf(
             'Illegal character in string found: %s at position: %d [%d:%d]',
             ord($char) < 32 ? ord($char) : $char,
             $position,
@@ -131,7 +134,7 @@ class ParseException extends JsonException
      */
     public static function illegalEscapeUsed(int $position, int $lineNumber, int $linePosition): ParseException
     {
-        return new static(\sprintf(
+        return new self(sprintf(
             'Illegal escape in string used at position: %d [%d:%d]',
             $position,
             $lineNumber,
@@ -148,7 +151,7 @@ class ParseException extends JsonException
      */
     public static function illegalOctalLiteral(int $position, int $lineNumber, int $linePosition, string $char): ParseException
     {
-        return new static(\sprintf(
+        return new self(sprintf(
             'Illegal octal literal: %s at position: %d [%d:%d]',
             $char,
             $position,
@@ -165,7 +168,7 @@ class ParseException extends JsonException
      */
     public static function illegalTrailingDecimal(int $position, int $lineNumber, int $linePosition): ParseException
     {
-        return new static(\sprintf(
+        return new self(sprintf(
             'Illegal trailing decimal at position: %d [%d:%d], a decimal point must be followed with ath least 1 digit',
             $position,
             $lineNumber,
@@ -181,7 +184,7 @@ class ParseException extends JsonException
      */
     public static function illegalNegativeSign(int $position, int $lineNumber, int $linePosition): ParseException
     {
-        return new static(\sprintf(
+        return new self(sprintf(
             'Illegal negative sign at position: %d [%d:%d], a negative sign may only precede numbers.',
             $position,
             $lineNumber,
@@ -197,7 +200,7 @@ class ParseException extends JsonException
      */
     public static function illegalEmptyExponent(int $position, int $lineNumber, int $linePosition): ParseException
     {
-        return new static(\sprintf(
+        return new self(sprintf(
             'Illegal empty exponent at position: %d [%d:%d], an exponent must be followed with ath least 1 digit.',
             $position,
             $lineNumber,
@@ -212,13 +215,13 @@ class ParseException extends JsonException
      */
     public static function unexpectedEndOfFile(Token $currentToken, array $expectedTokens): ParseException
     {
-        $plural = \count($expectedTokens) > 1;
+        $plural = count($expectedTokens) > 1;
 
-        return new static(\sprintf(
+        return new self(sprintf(
             'Unexpected EOF found. Last seen token: `%s`. Expected token %s: %s',
             $currentToken->getName(),
             $plural ? 'to be one of' : 'to be',
-            $plural ? '[' . \implode(', ', $expectedTokens) . ']' : $expectedTokens
+            $plural ? '[' . implode(', ', $expectedTokens) . ']' : $expectedTokens
         ));
     }
 
@@ -229,13 +232,13 @@ class ParseException extends JsonException
      */
     public static function unexpectedToken(Token $currentToken, array $expectedTokens): ParseException
     {
-        $plural = \count($expectedTokens) > 1;
+        $plural = count($expectedTokens) > 1;
 
-        return new static(\sprintf(
+        return new self(sprintf(
             'Unexpected token: `%s` found. Expected token %s: [%s]',
             $currentToken->getName(),
             $plural ? 'to be one of' : 'to be',
-            $plural ? '[' . \implode(', ', $expectedTokens) . ']' : $expectedTokens
+            $plural ? '[' . implode(', ', $expectedTokens) . ']' : $expectedTokens
         ));
     }
 
@@ -245,7 +248,7 @@ class ParseException extends JsonException
      */
     public static function illegalTrailingComma(Token $commaToken): ParseException
     {
-        return new static(\sprintf(
+        return new self(sprintf(
             'Illegal trailing comma found at position: %d [%d:%d]',
             $commaToken->getPosition(),
             $commaToken->getLineNumber(),
