@@ -7,13 +7,16 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed in the root of the source code
  */
+declare(strict_types=1);
+
 namespace Jojo1981\JsonAstBuilder\Lexer;
 
 use Jojo1981\JsonAstBuilder\Exception\ParseException;
+use UnexpectedValueException;
 
 /**
  * The interface for all lexer classes. There responsible for the lexical analysis and should produce a token stream.
- * When the input is not valid a a syntax error exception should be thrown.
+ * When the input is not valid a syntax error exception should be thrown.
  *
  * @package Jojo1981\JsonAstBuilder\Lexer
  */
@@ -21,8 +24,8 @@ interface LexerInterface
 {
     /**
      * @param string $input
-     * @throws ParseException
      * @return void
+     * @throws ParseException
      */
     public function setInput(string $input): void;
 
@@ -34,16 +37,16 @@ interface LexerInterface
     public function reset(): void;
 
     /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
      * @return Token
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     public function getCurrent(): Token;
 
     /**
-     * @throws ParseException
-     * @throws \UnexpectedValueException
      * @return Token
+     * @throws UnexpectedValueException
+     * @throws ParseException
      */
     public function getNext(): Token;
 }
